+++ src/CHANGES.md
@@ -1,8 +1,8 @@
 **Added:**
-* none
+* display properties editor
 
 **Changed:**
-* none
+* how a set macro is displayed in ControlPropertiesEditorPane (centered the label and made the macro key a hyperlink so that when it's clicked, opens macro editor popup)
 
 **Fixed:**
-* can't select more than one control style when creating a new control
+* none